api credentials moved to a separate file
adjusted panel opening

moved the title and handle to a separate child so that they're always on-top.
it has been done this way as a bandaid to mitigate the lack of a scrollController.

implemented glossary page + 7 days chart

- glossary page and details page
- wrote calculations for comparison examples
- implemented 7 days chart in home

started error handling

- custom error messages
- custom error classes
- first implementation

still missing a way to display the error to the user